# Improve perf of JS tests, fix proptype warnings