--- a/tests/js/spec/components/assigneeSelector.spec.jsx
+++ b/tests/js/spec/components/assigneeSelector.spec.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import TestUtils from 'react-addons-test-utils';
+import {mount} from 'enzyme';
 import AssigneeSelector from 'app/components/assigneeSelector';
 
 import LoadingIndicator from 'app/components/loadingIndicator';
@@ -75,7 +75,7 @@ describe('AssigneeSelector', function() {
         expect(putSessionUserFirst([USER_1, USER_2])).toEqual([USER_2, USER_1]);
       });
 
-      it('should return the same member list if the session user isn\'t present', function() {
+      it("should return the same member list if the session user isn't present", function() {
         this.sandbox.stub(ConfigStore, 'get').withArgs('user').returns({
           id: 555,
           name: 'Here Comes a New Challenger',
@@ -89,96 +89,86 @@ describe('AssigneeSelector', function() {
 
   describe('onFilterKeyDown()', function() {
     beforeEach(function() {
-      let assigneeSelector = (this.assigneeSelector = TestUtils.renderIntoDocument(
+      let assigneeSelector = (this.assigneeSelector = mount(
         <AssigneeSelector id="1337" />
       ));
 
-      this.sandbox.stub(assigneeSelector, 'assignTo');
+      this.assignTo = this.sandbox.stub(assigneeSelector.instance(), 'assignTo');
     });
 
     it('should assign the first filtered member when the Enter key is pressed and filter is truthy', function() {
       let assigneeSelector = this.assigneeSelector;
-      assigneeSelector.state.filter = 'Jane';
+      assigneeSelector.setState({filter: 'Jane'});
 
-      TestUtils.Simulate.keyDown(assigneeSelector.refs.filter, {
-        key: 'Enter',
-        keyCode: 13,
-        which: 13
-      });
-      expect(assigneeSelector.assignTo.calledOnce).toBeTruthy;
-      expect(assigneeSelector.assignTo.lastCall.args[0]).toHaveProperty(
-        'name',
-        'Jane Doe'
-      );
+      assigneeSelector
+        .ref('filter')
+        .simulate('keyDown', {key: 'Enter', keyCode: 13, which: 13});
+
+      expect(this.assignTo.calledOnce).toBeTruthy;
+      expect(this.assignTo.lastCall.args[0]).toHaveProperty('name', 'Jane Doe');
     });
 
     it('should do nothing when the Enter key is pressed, but filter is the empty string', function() {
       let assigneeSelector = this.assigneeSelector;
-      assigneeSelector.state.filter = '';
+      assigneeSelector.setState({filter: ''});
 
-      TestUtils.Simulate.keyDown(assigneeSelector.refs.filter, {
-        key: 'Enter',
-        keyCode: 13,
-        which: 13
-      });
-      expect(assigneeSelector.assignTo.notCalled).toBeTruthy;
+      assigneeSelector
+        .ref('filter')
+        .simulate('keyDown', {key: 'Enter', keyCode: 13, which: 13});
+
+      expect(this.assignTo.notCalled).toBeTruthy;
     });
 
     it('should do nothing if a non-Enter key is pressed', function() {
       let assigneeSelector = this.assigneeSelector;
-      assigneeSelector.state.filter = 'Jane';
+      assigneeSelector.setState({filter: 'Jane'});
 
-      TestUtils.Simulate.keyDown(assigneeSelector.refs.filter, {
-        key: 'h',
-        keyCode: 72,
-        which: 72
-      });
-      expect(assigneeSelector.assignTo.notCalled).toBeTruthy;
+      assigneeSelector
+        .ref('filter')
+        .simulate('keyDown', {key: 'h', keyCode: 72, which: 72});
+      expect(this.assignTo.notCalled).toBeTruthy;
     });
   });
 
   describe('onFilterKeyUp()', function() {
     beforeEach(function() {
-      this.assigneeSelector = TestUtils.renderIntoDocument(
-        <AssigneeSelector id="1337" />
-      );
+      this.assigneeSelector = mount(<AssigneeSelector id="1337" />);
     });
 
     it('should close the dropdown when keyup is triggered with the Escape key', function() {
       let assigneeSelector = this.assigneeSelector;
-      this.sandbox.stub(assigneeSelector.refs.dropdown, 'close');
+      let closeStub = this.sandbox.stub(
+        assigneeSelector.instance().refs.dropdown,
+        'close'
+      );
 
-      TestUtils.Simulate.keyUp(assigneeSelector.refs.filter, {key: 'Escape'});
+      assigneeSelector.ref('filter').simulate('keyUp', {key: 'Escape'});
 
-      expect(assigneeSelector.refs.dropdown.close.calledOnce).toBeTruthy;
+      expect(closeStub.calledOnce).toBeTruthy;
     });
 
     it('should update the local filter state if any other key is pressed', function() {
       let assigneeSelector = this.assigneeSelector;
 
-      TestUtils.Simulate.keyUp(assigneeSelector.refs.filter, {target: {value: 'foo'}});
-      expect(assigneeSelector.state.filter).toEqual('foo');
+      assigneeSelector.ref('filter').simulate('keyUp', {target: {value: 'foo'}});
+      expect(assigneeSelector.state('filter')).toEqual('foo');
     });
   });
 
   describe('componentDidUpdate()', function() {
     beforeEach(function() {
-      this.assigneeSelector = TestUtils.renderIntoDocument(
-        <AssigneeSelector id="1337" />
-      );
+      this.assigneeSelector = mount(<AssigneeSelector id="1337" />);
     });
 
-    it('should destroy old assignee tooltip and create a new assignee tooltip', function(
-      done
-    ) {
-      this.sandbox.spy(this.assigneeSelector, 'attachTooltips');
-      this.sandbox.spy(this.assigneeSelector, 'removeTooltips');
+    it('should destroy old assignee tooltip and create a new assignee tooltip', function() {
+      let instance = this.assigneeSelector.instance();
+      this.sandbox.spy(instance, 'attachTooltips');
+      this.sandbox.spy(instance, 'removeTooltips');
 
-      this.assigneeSelector.setState({assignedTo: USER_1}, () => {
-        expect(this.assigneeSelector.attachTooltips.calledOnce).toBeTruthy;
-        expect(this.assigneeSelector.removeTooltips.calledOnce).toBeTruthy;
-        done();
-      });
+      this.assigneeSelector.setState({assignedTo: USER_1});
+
+      expect(instance.attachTooltips.calledOnce).toBeTruthy;
+      expect(instance.removeTooltips.calledOnce).toBeTruthy;
     });
   });
 });