--- a/tests/js/spec/components/contextData.spec.jsx
+++ b/tests/js/spec/components/contextData.spec.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {mount} from 'enzyme';
+import {shallow} from 'enzyme';
 
 import ContextData from 'app/components/contextData';
 
@@ -8,7 +8,7 @@ describe('ContextData', function() {
     describe('strings', function() {
       it('should render urls w/ an additional <a> link', function() {
         const URL = 'https://example.org/foo/bar/';
-        let wrapper = mount(<ContextData data={URL} />);
+        let wrapper = shallow(<ContextData data={URL} />);
 
         expect(wrapper.find('span').at(0).text()).toEqual(URL);
         expect(wrapper.find('a').at(0).prop('href')).toEqual(URL);