--- a/tests/js/spec/components/userLetterAvatar.spec.jsx
+++ b/tests/js/spec/components/userLetterAvatar.spec.jsx
@@ -4,20 +4,20 @@ import UserLetterAvatar from 'app/components/userLetterAvatar';
 
 describe('LetterAvatar', function() {
   const USER_1 = {
-    id: 1,
+    id: '1',
     name: 'Jane Doe',
     email: 'janedoe@example.com'
   };
   const USER_2 = {
-    id: 2,
+    id: '2',
     email: 'johnsmith@example.com'
   };
   const USER_3 = {
-    id: 2,
+    id: '2',
     username: 'foo@example.com'
   };
   const USER_4 = {
-    id: 2
+    id: '2'
   };
   const USER_5 = {
     ip_address: '127.0.0.1'
@@ -72,7 +72,7 @@ describe('LetterAvatar', function() {
       this.letterAvatar = TestUtils.renderIntoDocument(
         <UserLetterAvatar user={USER_4} />
       );
-      expect(this.letterAvatar.getIdentifier()).toEqual(2);
+      expect(this.letterAvatar.getIdentifier()).toEqual('2');
     });
 
     it('should use ip address', function() {