@benvinegar 2017-05-01
Since this only affects tests, merging.